New Version

+++ libft/ft_atoi.c
@@ -6,16 +6,16 @@
 /*   By: fvargas <fvargas@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/07/17 23:23:13 by fvargas           #+#    #+#             */
-/*   Updated: 2024/07/18 14:05:41 by fvargas          ###   ########.fr       */
+/*   Updated: 2024/09/24 13:29:35 by fvargas          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_atoi(const char *nptr)
+long	ft_atoi(const char *nptr)
 {
-	int	sign;
-	int	result;
+	int		sign;
+	long	result;
 
 	sign = 1;
 	result = 0;
